wpt_babel/content-security-policy/object-src/object-src-no-url-blocked.html

<!DOCTYPE html>
<html>
 <head>
  <meta content="object-src 'none'; script-src 'self' 'unsafe-inline';" http-equiv="Content-Security-Policy"/>
  <script src="/resources/testharness.js">
  </script>
  <script src="/resources/testharnessreport.js">
  </script>
 </head>
 <body>
  <script>
   "use strict";

var t = async_test("Should block the object and fire a spv");
window.addEventListener('securitypolicyviolation', t.step_func_done(function (e) {
  assert_equals(e.violatedDirective, "object-src");
}));
  </script>
  <object type="application/x-webkit-test-netscape">
  </object>
 </body>
</html>
